refactor: moved cmp related fns to cmp.rs
