+++ src/daft-core/src/kernels/mod.rs
@@ -1,4 +1,5 @@
 pub mod binary;
+pub mod cmp;
 pub mod hashing;
 pub mod search_sorted;
 pub mod utf8;
